Python 3 support
This is probably still pretty sketchy and may bug out in parts of the Python code that haven't been exercised well...but it builds and runs and some basic smoke tests look ok. Only testing any of this on Windows.

Then again, updating a 12 year old project to a modern version of an embedded scripting engine and getting it to *actually do things* feels like a minor miracle even if some corner case ends up blowing up the interpreter at some point and needs to be fixed.