Add basic GMCP support
Completely untested in Perl and Ruby, but working in Python. The handler successfully manages to negotiate GMCP support with a server that implements it and processes incoming GMCP data by simply passing it to the script interpreter which then dumps it into a global dictionary (gmcpData) so that it's available to the user.

Improvements that probably need to be looked at:
 * Dealing with the server disabling/re-enabling GMCP (can this happen?)
 * Add some convenience functions in the interpreter. In particular something like sendGMCP() would be useful.
 * Raising some kind of event on new GMCP data, or some other kind of mechanism that will allow triggers to potentially process/react to them.